requirements.txt
fastapi
uvicorn
langchain
langchain-huggingface
langchain-core
langserve
python-dotenv
requests
huggingface_hub
streamlit
requests
